LayoutTests/fast/regions/webkit-region-rule-expected.txt

This tests the WebKitCSSRegionRule interface.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS WebKitCSSRegionRule exists on Window object
PASS WEBKIT_REGION_RULE exists on Window.CSSRule object
PASS regionRule1.type is window.CSSRule.WEBKIT_REGION_RULE
PASS regionRule1CSSRules.length is 2
PASS regionRule1CSSRules.item(0).type is window.CSSRule.STYLE_RULE
PASS regionRule1CSSRules.item(0).cssText is '#p1 { background-color: green; }'
PASS regionRule1CSSRules.item(1).cssText is '#p2 { background-color: lime; }'
PASS regionRule2.type is window.CSSRule.WEBKIT_REGION_RULE
PASS regionRule2CSSRules.length is 1
PASS regionRule2CSSRules.item(0).type is window.CSSRule.STYLE_RULE
PASS regionRule2CSSRules.item(0).cssText is 'p { background-color: blue; }'
PASS successfullyParsed is true

TEST COMPLETE
